Fix imports after module reorganization
- Updated test imports for new module structure
- Fixed source file imports in core modules
- CLI now imports from organized command modules
- All import paths updated to reflect new hierarchy
- Tests can now be collected and run successfully
