+++ src/llm_orc/core/execution/ensemble_execution.py
@@ -13,9 +13,9 @@ from llm_orc.core.execution.input_enhancer import InputEnhancer
 from llm_orc.core.models.model_factory import ModelFactory
 from llm_orc.core.config.ensemble_config import EnsembleConfig
 from llm_orc.models import ModelInterface
-from llm_orc.orchestration import Agent
+from llm_orc.core.execution.orchestration import Agent
 from llm_orc.core.config.roles import RoleDefinition
-from llm_orc.script_agent import ScriptAgent
+from llm_orc.agents.script_agent import ScriptAgent
 
 
 class EnsembleExecutor:
