+++ tests/test_enhanced_model_profiles.py
@@ -7,10 +7,10 @@ from unittest.mock import AsyncMock, patch
 import pytest
 import yaml
 
-from llm_orc.config import ConfigurationManager
-from llm_orc.ensemble_config import EnsembleConfig
-from llm_orc.ensemble_execution import EnsembleExecutor
-from llm_orc.roles import RoleDefinition
+from llm_orc.core.config.config_manager import ConfigurationManager
+from llm_orc.core.config.ensemble_config import EnsembleConfig
+from llm_orc.core.execution.ensemble_execution import EnsembleExecutor
+from llm_orc.core.config.roles import RoleDefinition
 
 
 class TestEnhancedModelProfiles:
