+++ tests/test_ensemble_execution.py
@@ -6,10 +6,10 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
-from llm_orc.ensemble_config import EnsembleConfig
-from llm_orc.ensemble_execution import EnsembleExecutor
+from llm_orc.core.config.ensemble_config import EnsembleConfig
+from llm_orc.core.execution.ensemble_execution import EnsembleExecutor
 from llm_orc.models import ClaudeCLIModel, ClaudeModel, ModelInterface, OAuthClaudeModel
-from llm_orc.roles import RoleDefinition
+from llm_orc.core.config.roles import RoleDefinition
 
 
 class TestEnsembleExecutor:
@@ -570,7 +570,7 @@ class TestEnsembleExecutor:
             )
 
             # Should fall back to Ollama
-            from llm_orc.models import OllamaModel
+            from llm_orc.models.ollama import OllamaModel
 
             assert isinstance(model, OllamaModel)
